requirements.txt
gensim==4.2.0
hdbscan==0.8.28
joblib==1.1.0
matplotlib==3.6.0
nltk==3.7
numpy==1.23.3
pandas==1.5.0
plotly==5.10.0
pymystem3==0.2.0
python-youtube==0.8.2
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
stop-words==2018.7.23
streamlit==1.13.0
tqdm==4.64.1
umap-learn==0.5.3
wordcloud==1.8.2.2
